requirements.txt
flask==3.1.0
google-generativeai==0.8.4
python-dotenv==1.0.1
pytest==8.3.4
